requirements.txt
SimpleWebSocketServer
websocket_client